+++ Assets/Resources/Scripts/Filipe/Cena Loja/CurrentMIssion.cs
@@ -3,11 +3,11 @@ using UnityEngine;
 public class CurrentMission : MonoBehaviour
 {
 
-    public float missionReward;
+    public int missionReward;
     public int RescueableAnimals;
     public int OperableAnimals;
     public bool End = false;
-    public float CurrentReward;
+    public int CurrentReward;
     public int Level;
     
     void Update()
